use wrapper function instead of variable for freebsd
This variable made me look various times to find if it was (e.g.)
needed to stub out the function in (unit)tests, but it looks like
it was only there for convenience (not having to define a function).

Let's change it to a regular function that wraps the underlying
implementation to prevent this confusion.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>