remove aliases for deprecated types and functions
These were added during the transition, but downstream projects now
use their non-deprecated variants. Unfortunately, BuildKit v0.21.0
shipped with an untagged version of go-archive, and depends on one
of these aliases, so we may need to wait for [buildkit@3f0a42e] to
be included in a v0.21.x patch release.

[buildkit@3f0a42e]: https://github.com/moby/buildkit/commit/3f0a42ece3fb19963bd83e1622111a27c9ff7d97

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>